commit du matin qui va bien

+++ arbre.cpp
@@ -21,11 +21,12 @@ Arbre::Arbre(Cellule* cell, int col, int row, const Essence* _essence, int _age,
 void Arbre::initialise()
 {
 	// Initialisation des points de vie de l'arbre
-	int rayon = essence->getDiametre() / 2;
-	int hauteur = essence->getHauteur();
+	float rayon = essence->getDiametre() / 2.0;
+	float
+	hauteur = essence->getHauteur();
 	
 	float masseV= essence->getMasse()/1000.0;
-	int points = (unsigned int) ( 4.0/3.0* (M_PI* (rayon*rayon)*hauteur) * masseV );
+	unsigned points = (unsigned int) ( 4.0/3.0* (M_PI* (rayon*rayon)*hauteur) * masseV );
 	
 	// Les données des essences sont décrites pour un arbre de 50 ans à 12% d'humidité
 	pv = points * age/50.0;
